backend/requirements.txt
# Core Backend Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database and Persistence
psycopg2-binary==2.9.10
alembic==1.16.4
SQLAlchemy==2.0.41

# Document Processing
PyPDF2==3.0.1
python-docx==0.8.11
pdfplumber==0.11.7
openpyxl==3.1.5
docx2txt==0.9
python-magic==0.4.27

# python-magic-bin==0.4.14 
markdown==3.5.1

# AI/ML and RAG Components
chromadb==0.5.23
sentence-transformers==3.0.1
langchain==0.2.17
langchain-community==0.2.19
langchain-text-splitters==0.2.4
numpy>=1.26.4,<2.0.0
pandas==2.3.1
scikit-learn==1.5.1

# Knowledge Graph
neo4j==5.25.0
py2neo==2024.1.5
spacy==3.8.7
langgraph==5.28.1

# LLM API Client
groq==0.30.0

# Voice Processing
openai-whisper==20231117
torch>=2.2.0
torchaudio>=2.2.0
scipy>=1.11.3

# Natural Language Processing
spacy==3.8.7
nltk==3.9.1
tiktoken==0.9.0

# File Handling and Utilities
aiofiles==23.2.0
pathlib2==2.3.7
requests==2.31.0

# Data Processing
lxml==6.0.0
pillow==11.3.0

# Development and Validation
pydantic==2.11.7
typing-extensions==4.14.1

# HTTP and Networking
httpx==0.28.1
aiohttp==3.12.14

# Additional Dependencies
PyYAML==6.0.2
click==8.1.8
tqdm==4.67.1


